Create utils.print.py file for pretty-print
